fix: rewrite tests for new code
