+++ package.json
@@ -44,6 +44,7 @@
     "eslint-plugin-react": "^7.28.0",
     "eslint-plugin-react-hooks": "^4.3.0",
     "jsdom": "^21.1.1",
+    "msw": "^1.2.1",
     "prettier": "^2.8.4",
     "sass": "^1.59.2",
     "typescript": "^4.9.3",
