+++ src/tests/Form.test.tsx
@@ -1,11 +1,17 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { it, describe, expect } from 'vitest';
+import { Provider } from 'react-redux';
 import FormAddCards from '../components/FormAddCard';
+import { store } from '../app/store';
 
 describe('Form', () => {
   it('render without crashing', async () => {
-    const { container } = render(<FormAddCards onSubmitData={() => {}} />);
+    const { container } = render(
+      <Provider store={store}>
+        <FormAddCards />
+      </Provider>
+    );
     expect(container.querySelector('.form')).toBeTruthy();
     expect(screen.getByLabelText('name')).toBeTruthy();
     expect(screen.getByLabelText('birthday')).toBeTruthy();
